# apps/api/openapi.yaml
openapi: "3.1.0"
info:
  title: ATHENA M&E Platform API
  version: "0.1.0"
  description: |
    Backend API for the ATHENA Monitoring & Evaluation platform.
    All endpoints (except /health and /auth/*) require a valid
    JWT issued via OAuth SSO, carried as an HttpOnly cookie
    (`athena_session`) or an `Authorization: Bearer <token>` header.
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://your.domain.com
    description: Production

# ── Security scheme ───────────────────────────────────────────
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: athena_session

  # ── Reusable schemas ────────────────────────────────────────
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    UUID:
      type: string
      format: uuid
    ISOTimestamp:
      type: string
      format: date-time
...
    Indicator:
      type: object
      properties:
        id: { $ref: "#/components/schemas/UUID" }
        org_id: { $ref: "#/components/schemas/UUID" }
        code: { type: string }
        name: { type: string }
        unit_of_measure: { type: string, nullable: true }
        baseline_value: { type: number, nullable: true }
        annual_target: { type: number, nullable: true }
        reporting_period_start: { type: string, format: date, nullable: true }
        reporting_period_end: { type: string, format: date, nullable: true }
        source_form_id: { $ref: "#/components/schemas/UUID", nullable: true }
        source_field_path: { type: string, nullable: true }
        aggregation_fn:
          type: string
          enum: [count, sum, avg, count_distinct]
          nullable: true
    Notification:
      type: object
      properties:
        id: { $ref: "#/components/schemas/UUID" }
        org_id: { $ref: "#/components/schemas/UUID" }
        user_id: { $ref: "#/components/schemas/UUID", nullable: true }
        type: { type: string }
        title: { type: string }
        body: { type: string, nullable: true }
        reference_id: { $ref: "#/components/schemas/UUID", nullable: true }
        reference_table: { type: string, nullable: true }
        is_read: { type: boolean }
        created_at: { $ref: "#/components/schemas/ISOTimestamp" }
    Pagination:
      type: object
      properties:
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
        pages: { type: integer }

  # ── Reusable responses ──────────────────────────────────────
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    UnprocessableEntity:
      description: Validation failed
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }

# ── Default security (override per-endpoint where needed) ─────
security:
  - sessionCookie: []

# ── Paths ─────────────────────────────────────────────────────
paths:
...
  /indicators:
    get:
      operationId: listIndicators
      summary: List all indicators with latest actuals
      tags: [Indicators]
      x-roles: [admin, supervisor]
      responses:
        "200":
          description: Indicators
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Indicator" }

    post:
      operationId: createIndicator
      summary: Create a SMART indicator
      tags: [Indicators]
      x-roles: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Indicator" }
      responses:
        "201":
          description: Indicator created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Indicator" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
  /indicators/{id}:
    get:
      operationId: getIndicator
      summary: Get indicator detail with all actuals
      tags: [Indicators]
      x-roles: [admin, supervisor]
      parameters:
        - name: id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/UUID" }
      responses:
        "200":
          description: Indicator with actuals
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/Indicator" }
                  - type: object
                    properties:
                      actuals:
                        type: array
                        items:
                          type: object
                          properties:
                            period_start: { type: string, format: date }
                            period_end: { type: string, format: date }
                            actual_value: { type: number }
                            computed_at:
                              { $ref: "#/components/schemas/ISOTimestamp" }
  /indicators/{id}/compute:
    post:
      operationId: computeIndicatorActuals
      summary: On-demand SQL aggregation — compute actual value for this indicator
      tags: [Indicators]
      x-roles: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/UUID" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [period_start, period_end]
              properties:
                period_start: { type: string, format: date }
                period_end: { type: string, format: date }
      responses:
        "200":
          description: Computed actual value
          content:
            application/json:
              schema:
                type: object
                properties:
                  actual_value: { type: number }
                  computed_at: { $ref: "#/components/schemas/ISOTimestamp" }

  # ── Reporting ─────────────────────────────────────────────────
...
  /notifications:
    get:
      operationId: listNotifications
      summary: Paginated notification feed for the current user
      tags: [Notifications]
      x-roles: [admin, supervisor]
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    {
                      type: array,
                      items: { $ref: "#/components/schemas/Notification" },
                    }
                  pagination: { $ref: "#/components/schemas/Pagination" }
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      summary: Mark a notification as read
      tags: [Notifications]
      x-roles: [admin, supervisor]
      parameters:
        - name: id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/UUID" }
      responses:
        "204": { description: Marked as read }
        "404": { $ref: "#/components/responses/NotFound" }